TSQ_MTC_SAR_RGB/requirements_2.txt
absl-py==2.1.0
bypy==1.8.5
cachetools==5.3.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
coloredlogs==15.0.1
cycler==0.11.0
dataclasses==0.6
dill==0.3.7
flatbuffers==23.5.26
fonttools==4.38.0
future==0.18.3
google-auth==2.26.2
google-auth-oauthlib==0.4.6
grpcio==1.60.0
humanfriendly==10.0
idna==3.6
importlib-metadata==6.7.0
joblib==1.3.2
keras==2.11.0
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mpmath==1.3.0
multiprocess==0.70.15
numpy==1.21.6
oauthlib==3.2.2
onnx==1.14.1
onnx-tool==0.9.0
onnxruntime==1.14.1
opencv-python==4.10.0.84
packaging==23.2
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sympy==1.10.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.4.5
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
typing_extensions==4.7.1
urllib3==1.26.18
Werkzeug==2.2.3
zipp==3.15.0
